backend/neuralhire-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
loguru>=0.7.0
groq>=0.4.0
openai>=1.12.0
elevenlabs>=1.10.0
mediapipe>=0.10.9
opencv-python>=4.9.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
sentence-transformers>=2.5.0
transformers>=4.38.0
torch>=2.2.0
pymupdf>=1.23.0
reportlab>=4.1.0
python-multipart>=0.0.9
resend>=2.0.0
pydantic[email]>=2.6.0
scipy>=1.12.0
passlib[bcrypt]>=1.7.4
